+++ server/config/config.go
@@ -35,10 +35,11 @@ type MysqlConfig struct {
 
 // RedisConfig defines configs related to Redis
 type RedisConfig struct {
-	Address  string
-	Password string
-	Database int
-	UseTLS   bool `yaml:"use_tls"`
+	Address          string
+	Password         string
+	Database         int
+	UseTLS           bool `yaml:"use_tls"`
+	DuplicateResults bool `yaml:"duplicate_results"`
 }
 
 const (
@@ -55,6 +56,7 @@ type ServerConfig struct {
 	TLS        bool
 	TLSProfile string // TODO #271 set `yaml:"tls_compatibility"`
 	URLPrefix  string `yaml:"url_prefix"`
+	Keepalive  bool   `yaml:"keepalive"`
 }
 
 // AuthConfig defines configs related to user authorization
@@ -139,9 +141,10 @@ type S3Config struct {
 
 // PubSubConfig defines configs the for Google PubSub logging plugin
 type PubSubConfig struct {
-	Project     string
-	StatusTopic string `yaml:"status_topic"`
-	ResultTopic string `yaml:"result_topic"`
+	Project       string
+	StatusTopic   string `yaml:"status_topic"`
+	ResultTopic   string `yaml:"result_topic"`
+	AddAttributes bool   `yaml:"add_attributes"`
 }
 
 // FilesystemConfig defines configs for the Filesystem logging plugin
@@ -211,6 +214,7 @@ func (man Manager) addConfigs() {
 	man.addConfigInt("redis.database", 0,
 		"Redis server database number")
 	man.addConfigBool("redis.use_tls", false, "Redis server enable TLS")
+	man.addConfigBool("redis.duplicate_results", false, "Duplicate Live Query results to another Redis channel")
 
 	// Server
 	man.addConfigString("server.address", "0.0.0.0:8080",
@@ -226,6 +230,8 @@ func (man Manager) addConfigs() {
 			TLSProfileModern, TLSProfileIntermediate))
 	man.addConfigString("server.url_prefix", "",
 		"URL prefix used on server and frontend endpoints")
+	man.addConfigBool("server.keepalive", true,
+		"Controls wether HTTP keep-alives are enabled.")
 
 	// Auth
 	man.addConfigString("auth.jwt_key", "",
@@ -325,6 +331,7 @@ func (man Manager) addConfigs() {
 	man.addConfigString("pubsub.project", "", "Google Cloud Project to use")
 	man.addConfigString("pubsub.status_topic", "", "PubSub topic for status logs")
 	man.addConfigString("pubsub.result_topic", "", "PubSub topic for result logs")
+	man.addConfigBool("pubsub.add_attributes", false, "Add PubSub attributes in addition to the message body")
 
 	// Filesystem
 	man.addConfigString("filesystem.status_log_file", "/tmp/osquery_status",
@@ -383,10 +390,11 @@ func (man Manager) LoadConfig() KolideConfig {
 			ConnMaxLifetime: man.getConfigInt("mysql.conn_max_lifetime"),
 		},
 		Redis: RedisConfig{
-			Address:  man.getConfigString("redis.address"),
-			Password: man.getConfigString("redis.password"),
-			Database: man.getConfigInt("redis.database"),
-			UseTLS:   man.getConfigBool("redis.use_tls"),
+			Address:          man.getConfigString("redis.address"),
+			Password:         man.getConfigString("redis.password"),
+			Database:         man.getConfigInt("redis.database"),
+			UseTLS:           man.getConfigBool("redis.use_tls"),
+			DuplicateResults: man.getConfigBool("redis.duplicate_results"),
 		},
 		Server: ServerConfig{
 			Address:    man.getConfigString("server.address"),
@@ -395,6 +403,7 @@ func (man Manager) LoadConfig() KolideConfig {
 			TLS:        man.getConfigBool("server.tls"),
 			TLSProfile: man.getConfigTLSProfile(),
 			URLPrefix:  man.getConfigString("server.url_prefix"),
+			Keepalive:  man.getConfigBool("server.keepalive"),
 		},
 		Auth: AuthConfig{
 			JwtKey:      man.getConfigString("auth.jwt_key"),
@@ -459,9 +468,10 @@ func (man Manager) LoadConfig() KolideConfig {
 			StsAssumeRoleArn: man.getConfigString("s3.sts_assume_role_arn"),
 		},
 		PubSub: PubSubConfig{
-			Project:     man.getConfigString("pubsub.project"),
-			StatusTopic: man.getConfigString("pubsub.status_topic"),
-			ResultTopic: man.getConfigString("pubsub.result_topic"),
+			Project:       man.getConfigString("pubsub.project"),
+			StatusTopic:   man.getConfigString("pubsub.status_topic"),
+			ResultTopic:   man.getConfigString("pubsub.result_topic"),
+			AddAttributes: man.getConfigBool("pubsub.add_attributes"),
 		},
 		Filesystem: FilesystemConfig{
 			StatusLogFile:        man.getConfigString("filesystem.status_log_file"),