+++ website/api/controllers/deliver-contact-form-message.js
@@ -23,15 +23,22 @@ module.exports = {
       example: 'I want to buy stuff.'
     },
 
-    fullName: {
+    firstName: {
       required: true,
       type: 'string',
-      description: 'The full name of the human sending this message.',
-      example: 'Hermione Granger'
+      description: 'The first name of the human sending this message.',
+      example: 'Emma'
+    },
+
+    lastName: {
+      required: true,
+      type: 'string',
+      description: 'The last name of the human sending this message.',
+      example: 'Watson'
     },
 
     message: {
-      required: true,
+      required: false,
       type: 'string',
       description: 'The custom message, in plain text.'
     }
@@ -48,32 +55,19 @@ module.exports = {
   },
 
 
-  fn: async function({emailAddress, topic, fullName, message}) {
+  fn: async function({emailAddress, topic, firstName, lastName, message}) {
 
-    if (!sails.config.custom.internalEmailAddress) {
+    if (!sails.config.custom.slackWebhookUrlForContactForm) {
       throw new Error(
-`Cannot deliver incoming message from contact form because there is no internal
-email address (\`sails.config.custom.internalEmailAddress\`) configured for this
-app.  To enable contact form emails, you'll need to add this missing setting to
-your custom config -- usually in \`config/custom.js\`, \`config/staging.js\`,
-\`config/production.js\`, or via system environment variables.`
+        'Message not delivered: slackWebhookUrlForContactForm needs to be configured in sails.config.custom. Here\'s the undelivered message: ' +
+        `Name: ${firstName + ' ' + lastName}, Email: ${emailAddress}, Topic: ${topic}, Message: ${message ? message : 'No message.'}`
       );
+    } else {
+      await sails.helpers.http.post(sails.config.custom.slackWebhookUrlForContactForm, {
+        text: `New contact form message: (Remember: we have to email back; can't just reply to this thread.) cc @sales `+
+        `Name: ${firstName + ' ' + lastName}, Email: ${emailAddress}, Topic: ${topic}, Message: ${message ? message : 'No message.'}`
+      });
     }
-
-    await sails.helpers.sendTemplateEmail.with({
-      to: sails.config.custom.internalEmailAddress,
-      subject: 'New contact form message',
-      template: 'internal/email-contact-form',
-      layout: false,
-      templateData: {
-        contactName: fullName,
-        contactEmail: emailAddress,
-        topic,
-        message,
-      }
-    });
-
   }
-
 
 };