+++ website/config/policies.js
@@ -23,5 +23,6 @@ module.exports.policies = {
   'legal/view-terms': true,
   'legal/view-privacy': true,
   'deliver-contact-form-message': true,
+  'view-query-detail': true,
 
 };